Update google2016 example
