+++ test_google2016_unbreakable/test.py
@@ -1,0 +1,78 @@
+def test():
+    from manticore.native import Manticore
+
+    """
+    Leverages Manticore to solve the Google 2016 challenge:
+    unbreakable-enterprise-product-activation
+
+    Author: @ctfhacker
+    Binary: https://github.com/ctfs/write-ups-2016/blob/master/google-ctf-2016/reverse/unbreakable-enterprise-product-activation-150/unbreakable-enterprise-product-activation.bz2
+    """
+
+    m = None
+
+    if __name__ == "__main__":
+        m = Manticore("unbreakable")
+    else:
+        m = Manticore("test_google2016_unbreakable/unbreakable")
+    m.context["solved"] = False
+
+    """
+    These values can be found at 0x4005b8
+    """
+    input_addr = 0x6042C0
+    num_bytes = 0x33
+
+    # Entry point
+    @m.hook(0x400729)
+    def hook(state):
+        """ CAUTION: HACK """
+        """ From entry point, jump directly to code performing check """
+
+        # Create a symbolic buffer for our input
+        buffer = state.new_symbolic_buffer(0x43)
+
+        # We are given in the challenge that the flag begins with CTF{
+        # So we can apply constraints to ensure that is true
+        state.constrain(buffer[0] == ord("C"))
+        state.constrain(buffer[1] == ord("T"))
+        state.constrain(buffer[2] == ord("F"))
+        state.constrain(buffer[3] == ord("{"))
+
+        # Store our symbolic input in the global buffer read by the check
+        state.cpu.write_bytes(input_addr, buffer)
+
+        # By default, `hook` doesn't patch the instruction, so we hop over
+        # the hooked strncpy and execute the next instruction
+        state.cpu.EIP = 0x4005BD
+
+    # Failure case
+    @m.hook(0x400850)
+    def hook(state):
+        """ Stop executing paths that reach the `failure` function"""
+        print("Invalid path.. abandoning")
+        state.abandon()
+
+    # Success case
+    @m.hook(0x400724)
+    def hook(state):
+        print("Hit the final state.. solving")
+
+        state._solver._command = "z3 -t:240000 -smt2 -in"  # Hack around travis timeouts
+        res = "".join(map(chr, state.solve_buffer(input_addr, num_bytes)))
+        print(res)  # CTF{0The1Quick2Brown3Fox4Jumped5Over6The7Lazy8Fox9}
+
+        with m.locked_context() as w:
+            if "CTF{0The1Quick2Brown3Fox4Jumped5Over6The7Lazy8Fox9}" == res:
+                w["solved"] = True
+
+        # We found the flag, no need to continue execution
+        m.kill()
+
+    m.should_profile = True
+    m.run()
+    assert m.context["solved"] == True
+
+
+if __name__ == "__main__":
+    test()
